requirements.txt
python-dotenv
vk-api